examples/advanced-pytorch/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytorch-example"
version = "1.0.0"
description = "Federated Learning with PyTorch and Flower (Advanced Example)"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.24.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "wandb==0.17.8",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "pytorch_example.server_app_secagg:app"
clientapp = "pytorch_example.client_app_secagg:app"

[tool.flwr.app.config]
num-server-rounds = 20
fraction-train = 0.25
fraction-evaluate = 0.5
local-epochs = 2
server-device = "cpu"
learning-rate = 0.01
num-shares = 3
reconstruction-threshold = 2
max-weight = 9000
timeout = 15.0

[tool.flwr.federations]
default = "local-sim"

[tool.flwr.federations.local-sim]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 2 # each ClientApp assumes to use 2CPUs
options.backend.client-resources.num-gpus = 0.0 # ratio of VRAM a ClientApp has access to
[tool.flwr.federations.local-sim-gpu]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.25
